Refactor ConnectionDetails controllers
